// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Rarity = {
  COMMON: 'COMMON',
  EPIC: 'EPIC',
  LEGENDARY: 'LEGENDARY'
} as const

export type Rarity = (typeof Rarity)[keyof typeof Rarity]


export const TradeStatus = {
  PENDING: 'PENDING',
  SENDED: 'SENDED',
  ACCEPTED: 'ACCEPTED',
  DECLINED: 'DECLINED',
  CANCELLED: 'CANCELLED'
} as const

export type TradeStatus = (typeof TradeStatus)[keyof typeof TradeStatus]


export const PackType = {
  COMMON: 'COMMON',
  EPIC: 'EPIC',
  LEGENDARY: 'LEGENDARY'
} as const

export type PackType = (typeof PackType)[keyof typeof PackType]


export const FriendshipStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED'
} as const

export type FriendshipStatus = (typeof FriendshipStatus)[keyof typeof FriendshipStatus]